# api.yaml
openapi: 3.1.0
info:
  title: Coaching Center System – MVP API
  version: 1.0.0
  description: |
    Phase-1 MVP backend APIs
    - Admin-driven system
    - Roles: ADMIN, TEACHER, STUDENT
    - Cookie-based JWT authentication
servers:
  - url: http://localhost:8000/api/v1
    description: Local server
security:
  - bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ============================
  # COMMON RESPONSES
  # ============================
  schemas:
    BaseSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
...
    CreateFeeRequest:
      type: object
      required: [batch_id, amount, period]
      properties:
        batch_id:
          type: string
          format: uuid
        amount:
          type: integer
        period:
          type: string
    CreatePaymentRequest:
      type: object
      required: [student_id, fee_id, amount_paid, payment_date, mode]
      properties:
        student_id:
          type: string
          format: uuid
        fee_id:
          type: string
          format: uuid
        amount_paid:
          type: integer
        payment_date:
          type: string
          format: date-time
        mode:
          type: string
...
paths:
...
  /admin/fees:
    post:
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeeRequest'
      responses:
        '200':
          description: Fee added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'

    get:
      tags: [Admin]
      responses:
        '200':
          description: List fees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
  /admin/payments:
    post:
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '200':
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'

    get:
      tags: [Admin]
      responses:
        '200':
          description: List payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
...